+++ src/file_header.h
@@ -1,3 +1,5 @@
+#include <stddef.h>
+
 /**
  *
  * @file file_header.h
@@ -31,7 +33,7 @@ void thirlage_init_file_header(thirlage_file_header *file_header, char *bytes);
 /**
  * @brief Initialize an empty *file_header using its *bytes with a page_size pages.
  */
-void thirlage_init_empty_file_header(thirlage_file_header *file_header, char *bytes, unsigned int page_size);
+void thirlage_init_empty_file_header(thirlage_file_header *file_header, char *bytes, size_t page_size);
 
 /**
  * @brief Validate *file_header, returning 1 if valid; 0 if not.
